refactored chart data creation into own helper method
